Better tiling.
Tiling on a flat wall is the same.

Tiling from side to side is the same

Tiling side to top or side to bottom has been adjusted so while it isn't
a perfect tile (can't be due to lighting adjustments), it is less
noticable due to major elements being lined up.

This technique will be used on future blocks where there are strong
visual elements combined with no obvious orientation (so far just
obsidian and bedrock.