Add build back in for 'release'
I'm going to leave the textures alone for a little while (may come back to it)
so putting the built version up so that its easier for people to see if this is
something they want to incorporate or extend.

If you like it and want to add, please submit a push request or similar.

If you like it and want to use it, if you use the png's just go ahead, some
attribution / link back here would be appropriate.

Or you can fork it as per most GPL projects and run with it under the same
(GPL) license terms.

As always the LICENSE supercedes the comment here and for any technical /
lawyer'ish questions I recommend contacting the EFF.  My intent however is that
distributing png's means a link back here and distributing an SVG means pushing
improvements back here.